PIM-4999: add pim date type and remove oro date type
