+++ src/Pim/Bundle/UIBundle/Resources/views/Form/pim-fields.html.twig
@@ -466,14 +466,14 @@
     {{ block('form_row_field') }}
 {% endblock %}
 
-{% block oro_date_widget %}
+{% block pim_date_widget %}
     {% set attr = attr|merge({ 'class': (attr.class is defined ? attr.class ~ ' ' : '') ~ 'datepicker', 'placeholder': placeholder|trans }) %}
     {% if form.vars.disabled %}
         {% set attr = attr|merge({'disabled': 'disabled'}) %}
     {% endif %}
 
     {{ block('date_widget') }}
-{% endblock oro_date_widget %}
+{% endblock pim_date_widget %}
 
 {% block pim_number_widget %}
     {{ block('number_widget') }}
